error correction

+++ tasks/mpi/karaseva_e_binaryimage/src/ops_mpi.cpp
@@ -198,8 +198,8 @@ bool karaseva_e_binaryimage_mpi::TestTaskMPI::RunImpl() {
   std::vector<int> neighbors;
 
   // Perform labeling for the local region assigned to the current process
-  int start_row = static_cast<int>(rank) * static_cast<int>(local_rows);
-  int end_row = static_cast<int>(rank + 1) * static_cast<int>(local_rows);
+  int start_row = rank * local_rows;
+  int end_row = (rank + 1) * local_rows;
   Labeling(input_, local_labeled_image_, rows, cols, 2, label_parent, start_row, end_row);
 
   // Ensure output buffer is allocated for gather
@@ -208,9 +208,8 @@ bool karaseva_e_binaryimage_mpi::TestTaskMPI::RunImpl() {
     return false;
   }
 
-  int result = MPI_Gather(local_labeled_image_.data(), static_cast<int>(local_rows * cols), MPI_INT,
-                          reinterpret_cast<int*>(task_data->outputs[0]), static_cast<int>(local_rows * cols), MPI_INT,
-                          0, MPI_COMM_WORLD);
+  int result = MPI_Gather(local_labeled_image_.data(), local_rows * cols, MPI_INT,
+                          reinterpret_cast<int*>(task_data->outputs[0]), local_rows * cols, MPI_INT, 0, MPI_COMM_WORLD);
 
   if (result != MPI_SUCCESS) {
     std::cerr << "[Rank " << rank << "] Error gathering labeled image data.\n";
